Initial work for IO and basic solving loop
A few key perspectives:
- The goal is to find an "interesting" conflict ASAP, not necessarily
find a SAT solution.
- We aren't building a (partial) assignment, we're building a trace of
actions which can be used to derive a (partial) assignment

Consequently the goal, essentially, is to find interesting conflicts
as quickly as possible, and if we fail to find a conflict we know it's
SAT. And if we can learn the [] clause, we're also done.

I am still unsure of the exact loop invariants (nevermind the loop
structure) entirely. So that's still a work in progress.

I am not sure how backtracking would really work in this case. For now
I think I will ultimately restart each time (each time learning a new clause).