Extremely primitive loop and tests established
It can now learn the clauses, and use that to get to the end of
iteration (reasonably) with the 8-clause unsat 3SAT test case (added).

There are lots of cases where we check the same sort of thing again
and again. There is also subtler redundancies like, if we instantly
learn the null-clause, we should instead mark that we're unsat.

Basically, cleaning up the code flow a bit. I think also probably
outlining "process" into the main loop would be reasonable too. I
don't want to get too ahead of myself with refactoring. I also want to
implement, e.g., smart clause learning and smart unit-prop (and then
smart backtracking, and all the other powerful features).