Hooked up initial (large-scale) tests, filled in more trace data
So far it looks like we're able to get the expected answers in the
usual satlib (small) inputs, but of course it takes a while. This is
good, though! I've hooked up a simple make-based test suite, tested
with minisat.

I also did a bit of work in sat.cpp to continue the plumbing for
getting all the conflict-reason stitching. In this way I hope to get
the "smarter" clause-learning up and running. Once I have those two
options, I'll test out my timing information and also see the delta
between our times and minisat's (expected to be *huge*... for now!).