Debugging watched literals, taking a step back before refactoring
There were some indexing (idx vs. c[idx]) bugs in watched literals,
but I believe I flushed those out. The big issue now is that many
assertions are failing. I think the root of it is that the control
flow is very involved.

I'm taking a check-point here (it passes the main correctness test
suite) and will try to refactor this. The big challenge is that we
should allow for multiple units in the queue, but after applying them
we may detect a conflict. At least, that's the latest I saw.