Migrated control flow to a state-based system
This is all, still, to help set the stage for apparently the thing I
struggle most with: implementing watched literals. I have a few shards
of changes here and there, but I'm still butting up against how to
handle unit clauses, empty clauses, and maintaining unit information
in the presence of backtracking.

On the one hand, unit clauses and empty clauses are "really easy" to
handle -- it's obvious what ultimately should be the end result of
finding one of those, at least descriptively. I really want to make
this code as simple as possible, so I want to really have to justify
special-casing those, though. For instance, my backtracking is robust
in the case of unit clauses (I think), and I want to test all the
different permutations of algorithm matches.

Speaking of which, there's an assertion failure (not sure if it
manifests as a solver failure) under one combination -- see what's set
in the makefile.