Continuing prep for watched literals; bugfix
This is the work I did last night. If I want the invariant that
watched literals are only for clauses of size greater-than-1, and that
when we add a new clause to the watched literals it has at least one
not-false literal (both of these seem natural in my exploratory
efforts, at least), that suggests our main driver loop should
explicitly take the learned clause, think about it (backtracking), and
only after backtracking actually insert it in the CNF. So I added that
loop logic.

I also got caught up investigating one of the combinations of
algorithms that fails. I finally realized that even in the simplest
backtracking, we don't want to throw away the unit-prop actions from
the unit clauses. So I added that logic, and that's that. In the
makefile I also changed things so I can easily test all the algorithm
combinations I'd like.

I lose the ability to easily measure optimization effects, but I did
so "manually" and it is really neat to see how things speed up for
benchmark2.

Lastly, I isolated a lot of the correctness asserts and wrapped them
in a macro. When I don't define it, things get a *lot* faster. I think
the fastest benchmark2 run goes from about 12 seconds to 4 seconds,
single-threaded, on my machine. Nothing scientific, but it's pretty
clear (and not too surprising, to be honest). To suppose the ease of
making it debug-only (default on), I put some correctness-checking
logic in their own functions so I can just do SAT_ASSERT(checker).

Still slowly converging on an actual watched-literals
implementation. I think ironing out how the unit queue works (and
building off of that machinery with the watched literal clauses), and
having the piping to support *not* adding length-1 or length-0 to the
watched literals machinery, will help make this successful.