Implemented a correct watched literals test
I think so, at least. I have some pretty intense validation that seems
to pass and meet most of my expectations. Unfortunately, it seems that
even in benchmark 3, with the -j option (so maybe that's somehow
clouding things?) it takes 45 seconds with watched literals *or* with
the more naive queueing system.

Still, it works! And maybe there will be some natural extensions to
make it better. Later.