Green correctness tests; literal_incidence_map
This is a long-awaited refactor and re-enabling of a lot of the
correctness checks I've already written. Thankfully nothing jumped out.
Before I continue to consume perf information to improve the fundamental
data structures, I need to make sure it's accessible to the last
fundamental operation: removing clauses in the middle of execution.

My brainwave, such that it is, is to try to apply BCE after learning
lots of clauses. So long as that finds some clauses to remove, and
removes them, that should be enough to exercise the fundamental plumbing
I'm interested in.

The refactoring I've been doing has been highlighting an interesting
fact: I *think* that only the watched literals machinery has state that
depends on the state of the CNF. Learning, backtracking, and
learned-clause-minimization do not. They are instead dependent on a
trail (which implicitly cares about the state of the CNF, as it indexes
into it). That's interesting! Something to think about.

It suggests that clause-cleaning shouldn't be *too* bad -- just making
things consistent with the trail and the watched literal machinery.
We'll see.