VSIDS; data structure refactoring
I generalized the literal-to-mapping used, and applied it to VSIDS as
well. That is a really impressive speedup, over halving the benchmark3
time. We're now down to about 8 seconds. Minisat still manages it in 2.2
or so, so there's still lots to do.

There are definitely some additional data structure improvements on the
table. One I'm interested in is getting rid of the std::map in
watched_literals. I generally want to avoid doing these smaller
implementation optimizations, except that I think it will help pave the
way for enabling sane in-processing clause-removal. You can see some
initial attempts in cnf_t. The idea is that each clause is assigned an
(extremely long-term) clause-key, and *that* keys into an array of the
actual memory. So we can move that backing memory around more easily.
More layers of indirection...

Anyways, a clause-key-to-data map is exactly what we use in
watched_literals as well, so I might as well start there.