Added clause_key_map (clause_map_t) to TWL
This leads to a 7.2->6.5 second speedup on benchmark3, for instance. So
that's nice. A bit touch-and-go with making sure the right size fields
are added in the right order, but I'm comfortable with the correctness
here so far.

The next step is to integrate this in the cnf object to enable clause
removal without having to rekey everything.