Simple unit queue
Small project as a diversion from the real next hurdle, which is to
create a CNF structure that supports clause removal efficiently (without
having to rekey everything). This is basically just a not-quite-circular
queue. It passes all tests under assertions, so I thought to save it at
this point. Right after I'll see if it actually leads to a speed
improvement.

This is part of the other project of moving further and further away
from complicated data structures (where possible).