LBM, counters, and LCM investigation
I did a first, very expensive, pass of LBM, and I still get a 20%
speedup in benchmark3 (going from 6 seconds to 5 seconds).

I implemented some basic counters to sanity-check against minisat...
there are many more conflicts, much larger clauses, and all that
relative to minisat. This is surprising! I suspect that I may not be
doing recursive learned-clause minimization correctly, but we'll see. I
am assuming this is indicative of a real problem in my solver.

In the meantime, I found the SAT 2019 race papers, and they basically
all cited that preprocessing paper from 2005, so I guess that's the next
new machinery on the list to implement.

I'm pleased with how things are going now. There's also other methods of
learned-clause minimization I want to explore.