Saving some experiments with subsumption
It plays a large role in our learned clauses for 3SAT, because they're
enormous, but doesn't play a large role in the initial
preprocessing. I'm concerned that this is indicative that I've not
implemented learned-clause-minimization correctly, but even after a
few variations it does seem consistent.

I also added some counters and similar.