Initial plugins framework
I got caught by some issues (in the now-addressed assertions) because I
wasn't maintaining all the information correctly when removing clauses.
I figured that was good enough motivation to implement a plugins system,
so I did.

My impression is that std::function is not the cheapest machinery, but
for now I still get about 5.2 seconds on benchmark 3, which I believe is
not quite as fast as 5.0 I got at one point, but I actually had that
just before implementing plugins. I think maybe now that I'm using the
last std::map (in the literal_to_clauses field) that is the new
overhead.

Either way, I think removing that dependence is the next thing.