More general plugin model; start of visualizations
I want to be able to visualize what the solver is doing. As one
example I am trying to get it to emit a tikz rendering of the
implication graph during a conflict. The motivating example (the first
output of "make fix" seems to be taking an extremely long time to
render. May need to take a slightly more hands-on approach and provide
more structure, if the issue is that the layout algorithm is overwhelmed.