Developing a harness to test clause learning
I can read in a trail, and perform the *same* clause-learning that we
don in the solver, but as a stand-alone. So I can compare things to what
minisat is doing, for instance, with just some fairly-blind extensions
to minisat.

I enhanced the trail so that the whole trail is consistent, and I don't
need to load in the whole CNF. So it is something of a minimum repro! I
have an example where we don't learn as much as minisat. I can either
analyze that, or develop more of a framework to find the smallest
example where we don't learn as much. Maybe the latter would be wiser.