One fix to LCM
The issue is that we mark *everything* in the clause, even decision
variables. This makes a lot of sense and was an obvious bug, in my
eyes. So I removed the "marked" vector entirely and instead just
containment in the clause.

There is still another issue, not quite sure what yet. Investigating,
but this is already expected to be an improvement: it's green for the
test3, and I need to go against the (much larger) test4 to find
another repro case. This one's even more exotic.