First round of perf investment
Callgrind made it clear that we spend an enormous amount of time
updating the literal_incidence_map of literals to clause_ids while doing
unit prop in watched_literals.

I tried to make that the obvious just-dynamic-array-thing, but I made
dumb mistakes, so I wrapped a vector (and started wrapping things as
pointers, instead of bigger iterators).

This already has the uuf250-01.cnf from 60 seconds to 50 seconds, or so.
Very happy. Next up is to make it actually-just-an-array.