Improved watched literals
Simple reasoning extension to watched literals to greatly improve
peformance. The uuf250-01.cnf benchmark goes from about 45-50 seconds to
about 35 seconds. This was informed by a cachegrind trace (because perf
doesn't work on WSL yet...), as was the previous effort. To be honest,
the most it revealed was that there was a lot of time spent in this
function, which is sort of a known aspect of CDCL with watched literals
anyways.

Minisat can still solve the benchmark in *3* seconds. So there is still
a long way to go.

My efforts at moving things away from vectors and into simpler data
structures actually didn't work out -- things seemed to get slower?

I am very hesitant to "embed" the watched literals as the first/second
literals in the clause. There is a lot of advantage, I find, to keeping
the clauses immutable, at least conceptually. Maybe I will revisit that,
though.