Small improvements from callgrind
Down to about 20 seconds. Still not close to the minisat 3-4 seconds on
this benchmark. Callgrind gives a fairly high granularity of where the
time is actually spent, but I suspect cache issues are the major
bottleneck. How might I confirm this?

Will try to measure cache hits with cachegrind and see how it turns out.