Improved LCM
I noticed that on the current profile run (200 vars) about 15% of the
time was in LCM, 10% of which was in LCM proper. Looking at the tracing,
it is that we spend a lot of time looking for the action that induces
the unit-prop in the trail. Now we cache those, and in the following
profile LCM goes to about 12% of the time, with LCM proper taking only
6%.

There is still improvement to be made here, I have ideas. This is much
more tractable than my various failed experiments with improving TWL.