Perf improvements cont'd
The VSIDS changes here don't register an improvement, but I can only
hope it will be an investment for when I finally get around to a better
scaling algorithm.

The LBM improvement is considerable: it goes from about 8.1 seconds to
7.5 seconds, which is about an 8% improvement or so.

On this particular example, minisat takes about 4.3 seconds. So we're
closing in!