Major refactor: moved solver to solver object
There is now a discrete "solver" object that has its own copies of all
data (except for a few leftover static variables in LCM and
clause-learning and the like, which I will address soon).

The motivating benchmark remains at about 8 seconds or so. Still work to
do, but this is a much "cleaner" starting point.