Adding vivification preprocessing and some cleanup
There are some other changes I haven't included here. The main work with
vivification was crafting it as if it were a solver. I'm actually quite
pleased with how it turned out -- the only issue is that it doesn't
actually speed anything up, and in fact seems to slow things down! It
also rarely hits in the, e.g., 250 benchmark. So I'll be looking at this
more.

I think I may want to try to get restarts back into the game, and really
(finally) revisit the proper way of choosing literals and restart
strategies.