Misc experiments and improvements
I tried hacky implementations of ACIDS and VMTF, but I couldn't get VMTF
correct and ACIDS was a whole lot slower (albeit I didn't try that
caching trick). VMTF also motivated disentangling backtracking a bit,
which I was happy with.

I also tried to isolate some measurement code in ... "measurements". I
also tried a few variations of LCM, thinking maybe there was an
opportunity there -- no luck so far.

I'm sure there is still a perf gap to account for with minisat, but I'm
inclined at this point to consider algorithmic improvements. I don't
want to hide performance opportunities, though, that's the concern.

Right now on ./profile, TWL takes about 50% of the time. It should be
80%, is my understanding. I wonder how much overhead my plugins are
causing. Maybe that's something to account for. There is also counting
how many decisions and propogations to make, and making sure we're
"algorithmically" competitive with minisat. Lots of things to explore.