More cleanup; real improvement with LCM
I changed it so that we don't call "contains", but instead precompute a
mapping of literals to whether they're in the clause.

Other misc cleanup. I've found that there is an issue with
remove_clause_set, as LBM doesn't actually implement that, and I'm not
sure how best to implement that efficiently... ironic.