Implemented an experiment for a more cache-friendly clause type
Perf analysis is somewhat limited due to WSL environment (I think).
Not many counters available, but it got page faults. On uuf250-01.cnf,
the vector sort of page faults typically hovered around 5200, but with
this change it now goes to about 4200 to 4700. So 10-20% reduction.

This is good! And it seemed to scale roughly linearly with runtime, which
is as expected. I will also revisit the TWL implementation (again...) as
the data structure makes it plain that keeping active literals in the
hot memory can be very useful. And then I will revisit the rest of the
memory style.