Using faster TWL; improved with hand-written slide
Really building off the idea of caches, and saving literals that had a
lot of activity. This does seem to be a real perf improvement on
uuf250-01.cnf, but there's still a gap to go. Minisat has 500 or so page
faults, I still have over 4000.

I think the next thing is to really revisit how we're allocating
clauses, and avoiding a lot of stupid copies. I know that's been around
for a while, and now it's a long-pole.