Moving to different clause model, clang-format
Now clause_ids are really just pointers directly to clauses. This
reduces our page faults to about 1800, which is getting a lot closer to
the 400 or so in minisat.

However, we're actually still hovering around the same time. Why? I
would guess (haven't looked yet) that it's because I have to sort things
differently.

It's becoming clear that "cnf_t" should really just be a clause
allocator, and that clauses should be in a doubly-linked list. We'll
see. I previously trusted that it was advantageous to keep all clauses
in a flat array -- maybe not so, anymore.