Implemented "sliding" list deletion
Went back to value-based clauses with no real time loss. Changed
watched-literals iteration method so that I'm not always comparing
against the end, but instead waiting for the pointer to be 0. That got a
70->62 seconds speedup on benchmark 9, or there-abouts.

From that same mindset I changed watched literals so we're not
constantly swapping with the end, but instead "sliding down". This got
us from 62->50 seconds on benchmark9. That's huge! I'm wondering if
there are other opportunities, now.