Update-cause clause experiments, negative result
When we try to add a new literal, cause pair to the unit list, sometimes
that's redundant. We can see if the new cause is "better" than the
current cause we have. Currently the experiment was to see if it's
smaller. However, this does not lead to a performance boost. It's not
clear why, but my hypothesis is that we're having the effect of
lengthening the resolution chain. I started to install some metrics.

Still at about 50seconds for benchmark9. Not sure how to close the gap.