VSIDS restarting finally gets us there
Benchmark9 in about 32 seconds! My goodness. That was a lot, that's a
huge jump. I hadn't appreciated that I needed to (essentially) reset the
vsids counter on every reset, but once the thought occurred to me it
made it clear.

This is an enormous win, that's like 48 seconds to 32: 33% or 50% or
whatever improvement. What measurements could have informed this? I'm
not sure. Something to think about.

Now let's see if we're really competitive on all these inputs with
minisat.