Conflict clause histogram; exploring state ideas
I decided to see how frequently a particular clause was the motivating
conflict. Sure enough, it's almost a perfect curve. I have some simple
unix commands (sort | unique -c, and then | sort -nr for sort-by-amount)
to help do the statistics.

The perfect curve, I mean it looks very much like there are a few
high-conflict clauses, then it quickly goes into a long tail. Not quite
as sharp as you might hope for, but pretty darn sharp.

Is this an artifact of how I'm determining when a conflict occurs? (As
in, are there "multiple" clauses involved and I'm only counting one, in
some sense?) Even if this is completely correct, how to make this
information actionable? The heavy-hitters in random-3-sat are 3-clauses,
not the (much larger) learned clauses.

By hand, it looks like the heavy-hitter clauses appear at "random"
times, it's not like a signal showing convergence.

Say this is all correct: how to improve the search if we're already down
to a 3-CNF? Can we have to vivify it into a 2-clause, or better?