Soundly beat minisat with better VSIDS; trace CDCL
The perf improvement is from the (very small!) edit to installing VSIDS,
but it means we're suddently substantively better than minisat on our
benchmarks. Holy crap!

This came about from the new "plugin" object, cdcl_resolve, which made
it easy to add in the new literal-bumping behavior. Weirdly I had tried
to approximate this manually before and didn't see a speedup; I think I
had "bumped" more literals than I was supposed to in that case.

Anyways, now there's new flag-parsing so you can pass in a string as
part of a flag. Definitely ad-hoc for now, but it helped me produce the
clauses involved in the CDCL for the most commonly-conflicted clause.
It, uh, hasn't been super-illuminating. I suppose my long-term goal
there would be try to "flatten" out which clauses are most conflicted,
from the assumption that helps expedite the solution search space, but
I'm really just exploring to see if something jumps out.