Added viv preprocessor
This has a completely separate machinery from the "normal" solver,
though there are obvious echos. This is to keep avenues open for
different inprocessing algorithms. We can either "piggy-back" off of the
solver's main watched literals and unit queue, or we can do it
ourselves. Would like to keep things separate.

Big fix was turning off assertions that were accidentally always
running. This now gives a nice 3% win or so on the cnfsearch benchmarks,
31 seconds to 30 seconds or so (pretty reliably). That's just for
preprocessing.

Not sure how to validate that we really are catching all vivification
opportunities. We're sound, as far as I can tell, but complete? Hrm.