Setting stage for rest of on-the-fly subs
A big refactoring of where the plugins are called and such. This is
turning the clause-learning function into a "main" function that's
expected to call various plugins, versus simply a component of another
plugin. I think this is ultimately the right thing.

The big functional change is that the clause learning now returns the
clause_id, rather than the clause value (so we add it to the CNF
earlier). This is so that in the on-the-fly case we can return the
existing clause "transparently".

Note the previous experiments confirmed:
- Even in the case where our "learned" clause is a subsumption of an
existing clause, there are times when we can still do LCM. The trail
itself is telling us more useful information.
- That in turn suggests an experiment: try LCM *every* time there's a
subsumption hit, on-the-fly.
- There are times when we on-the-fly subsumption into a *unit* clause.

Separately, turning off LCM tripped an assertion in backtracking.
Experimentally I confirmed that it was capturing the case when we didn't
derive the empty clause, but instead backtracking went "all the way
back" and we should report this unsatisfiable.