Wrapping up OTFS
Now we can handle any counter. Experimentally validated that so far we
aren't shrinking anything except the learned clause to unit size, though
I don't have a rigorous conclusion.

Some technical debt in how we handle the integration of LCM, having to
"suspend" a clause and then restore it; something to revisit with a more
principled design.

Perf gains are real (motivating benchmark now 28 seconds versus 31, so
about a 10% win, in cnfsearch) but not as dramatic as I'd like. Still,
good to have this algorithm done, and forcing my hand at a more generous
implementation of CDCL and LCM.