Wrote up docs
Getting ready to make this public. Writing up an annotated bibliography
of about maybe 2/3rds of the papers I've poked at was a really valuable
exercise, actually.

Removed the MSVC build scripts; they're simply not maintained. Whoops.
Brought the README to, I think, reflect reality correctly.

Maybe after a bit more work this will be ready to be put to rest again
for the hibernation period.