Have stamped reflect current resolvent
Something I've circled is that the "stamped" set in the clause learning
routine *is* the current resolvent. (Well, prior to this change, it was
all the literals, negated.) So I made that change (validated offline) to
confirm that it really is that interior resolvent.

I have a commented-out experiment where I continue the on-the-fly (OTF)
subsumption even after learning the clause. This caused invariant
violations, where backtracking leaves a newly-subsumed clause to be unit
(where previously, I guess, it would have had 2 unset literals). Not
sure yet what to do about this; tempted to consult literature.

Left in a few tools and (commented-out) checks to that end.