Major refactor, change to clause-learning
I did a number of perf experiments. Shifting to the B&T integer set had
a small (10%) slowdown in the main benchmark; the biggest thing was that
depending on where the associated set was declared in the object, we had
an even bigger slowdown! That motivated the PGO builds, which didn't
work to remove that sort of noise.

This enables us to treat the "stamped" set as if it were exactly the
resolvent at each step in the trail. I think that's handy and neat.